Move `modal.tsx` to folder
